requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
packaging==24.0
PyJWT==2.8.0
sqlparse==0.5.0
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.3
